remove front unused code
