--- frontend/src/app/components/feature-card.tsx
@@ -1,21 +1,0 @@
-import type { LucideIcon } from "lucide-react";
-
-interface FeatureCardProps {
-  title: string;
-  description: string;
-  icon: LucideIcon;
-}
-
-export function FeatureCard({
-  title,
-  description,
-  icon: Icon,
-}: FeatureCardProps) {
-  return (
-    <div className="bg-gray-800 p-6 rounded-lg shadow-lg transition-all duration-300 hover:scale-105 hover:bg-gray-700">
-      <Icon className="w-12 h-12 text-primary mb-4" />
-      <h3 className="text-xl font-semibold mb-2">{title}</h3>
-      <p className="text-gray-300">{description}</p>
-    </div>
-  );
-}
