--- frontend/src/app/components/ui/badge.tsx
@@ -1,35 +1,0 @@
-"use client";
-
-import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "../../lib/utils";
-
-const badgeVariants = cva(
-  "inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium",
-  {
-    variants: {
-      variant: {
-        default: "bg-gray-100 text-gray-800",
-        primary: "bg-blue-100 text-blue-800",
-        secondary: "bg-gray-100 text-gray-800",
-        success: "bg-green-100 text-green-800",
-        destructive: "bg-red-100 text-red-800",
-        warning: "bg-yellow-100 text-yellow-800",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-    },
-  },
-);
-
-interface BadgeProps
-  extends React.HTMLAttributes<HTMLDivElement>,
-    VariantProps<typeof badgeVariants> {}
-
-export function Badge({ className, variant, ...props }: BadgeProps) {
-  return (
-    <div className={cn(badgeVariants({ variant }), className)} {...props} />
-  );
-}
-
-export { badgeVariants };
