--- frontend/src/app/types/wallet.ts
@@ -1,12 +1,0 @@
-export interface Transaction {
-  id: string;
-  type: "send" | "receive";
-  amount: string;
-  date: string;
-}
-
-export interface WalletData {
-  address: string;
-  balance: string;
-  transactions: Transaction[];
-}
